put mood db population program into a separate file populateMoodDB.js
